Add Git configuration

+++ install.sh
@@ -6,11 +6,21 @@ SCRIPT_DIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" &>/dev/null && pwd -P)
 source $SCRIPT_DIR/zsh/modules/helper/init.sh
 
 cd ~/ || exit
-# https://github.com/git/git/blob/master/Documentation/RelNotes/1.7.12.txt#L21-L23
-mkdir -p .config/git
-ln -fs "$SCRIPT_DIR/.gitignore_global" .config/git/ignore
 ln -fs "$SCRIPT_DIR/.tmux.conf" .
 ln -fs "$SCRIPT_DIR/.vimrc" .
+
+#
+# Git
+#
+cd ~/ || exit
+# https://github.com/git/git/blob/master/Documentation/RelNotes/1.7.12.txt#L21-L23
+mkdir -p .config/git
+if [ ! -f .config/git/ignore ]; then
+  ln -fs "$SCRIPT_DIR/.gitignore_global" .config/git/ignore
+fi
+if [ ! -f .gitconfig ]; then
+  ln -fs "$SCRIPT_DIR/.gitconfig" .
+fi
 
 # Copy settings.json if this script is running in a container
 if is-docker; then
